backend/monitoring/telemetry.py

"""
@file backend/monitoring/telemetry.py
@description Prometheus, Android Mobile, and eBPF telemetry hooks.
"""

import asyncio
from prometheus_client import start_http_server, Gauge, Info
import psutil
import time

# Resource Gauges
GPU_MEM = Gauge('eudorax_gpu_memory_usage_bytes', 'Total GPU VRAM in use')
CPU_HUGEPAGES = Gauge('eudorax_hugepages_free', 'Free Transparent Hugepages on system')
EBPF_SCHEDULER_LATENCY = Gauge('eudorax_ebpf_scheduler_ns', 'Kernel scheduler latency in nanoseconds')
ANDROID_NPU_TEMP = Gauge('eudorax_mobile_npu_celsius', 'Android Device NPU Temperature')
ENGINE_INFO = Info('eudorax_engine_build', 'System version information')

def init_prometheus(app):
    """Initializes the telemetry loop."""
    ENGINE_INFO.info({
        'version': '4.5.0-2026',
        'compiler': 'Clang 19 (AVX-512)',
        'backend': 'Python 3.14 (Astral)',
        'cuda': '13.1',
        'bmm3': 'enabled',
        'ebpf': 'active'
    })
    
    # Simple background task for resource tracking
    import threading
    def track():
        while True:
            # Simulation: In production use pynvml and bcc for eBPF metrics
            GPU_MEM.set(18 * 1024 * 1024 * 1024) 
            CPU_HUGEPAGES.set(1024)
            EBPF_SCHEDULER_LATENCY.set(120) # 120ns
            threading.Event().wait(10)
            
    threading.Thread(target=track, daemon=True).start()

async def android_telemetry_loop():
    """Simulates a loop gathering data from connected Android devices."""
    while True:
        # Monitoring Android NPU and sensors
        ANDROID_NPU_TEMP.set(42.5) # Simulated 42.5°C
        await asyncio.sleep(5)
