backend/requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.30.0
uvloop==0.21.0
httptools==0.6.1
firebase-admin==6.5.0
prometheus-client==0.20.0
polars==1.0.0
pyarrow==17.0.0
qdrant-client==1.10.0
torch==2.6.0+cu131
cupy-cuda12x==13.1.0
numba==0.60.0
transformers==4.44.0
tiktoken==0.7.0
scipy==1.14.0
opencv-python-headless==4.10.0.84
real-esrgan==0.3.0
clip-anywhere==0.2.1
pydantic==2.8.2
argparse
pytest==8.0.0
pytest-asyncio==0.23.5
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
